1) Remove the namespace from Module 1. 2) Make Module 1 usable in main.py. - Deleted the projects scripts file science replaced with Modules file - Set the main.py in the rout directory where it should be - Kept the logging after making sure it worked - Did an example of how should we do points 1&2 for the rest modules and there's TODO statement for each module.
