--- ProjectScripts/Modules/module3_VV_LOVD/.idea/inspectionProfiles/Project_Default.xml
@@ -1,28 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="8">
-            <item index="0" class="java.lang.String" itemvalue="vv_flask_restful_swagger" />
-            <item index="1" class="java.lang.String" itemvalue="VariantValidator" />
-            <item index="2" class="java.lang.String" itemvalue="vvhgvs" />
-            <item index="3" class="java.lang.String" itemvalue="VariantFormatter" />
-            <item index="4" class="java.lang.String" itemvalue="flask-log" />
-            <item index="5" class="java.lang.String" itemvalue="flask_restx" />
-            <item index="6" class="java.lang.String" itemvalue="flask" />
-            <item index="7" class="java.lang.String" itemvalue="dicttoxml" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-    <inspection_tool class="PyUnresolvedReferencesInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredIdentifiers">
-        <list>
-          <option value="application.app_v1.restx" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
