--- ProjectScripts/Modules/module3_VV_LOVD/docker.ini
@@ -1,44 +1,0 @@
-[mysql]
-host = vdb
-database = validator
-user = vvadmin
-password = var1ant
-
-[seqrepo]
-version = 2018-08-21
-location = /usr/local/share/seqrepo
-
-[postgres]
-host = uta
-database = uta
-version = uta_20171026
-user = anonymous
-password =
-
-[logging]
-#Levels control verbosity and can be set to "CRITICAL" "ERROR" "WARNING" "INFO" or "DEBUG".
-log = True
-console = ERROR
-file = ERROR
-
-[Entrez]
-email = <Your Email Address>
-api_key = <Your API Key>
-
-
-# <LICENSE>
-# Copyright (C) 2019 VariantValidator Contributors
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-# </LICENSE>
