--- ProjectScripts/Modules/module3_VV_LOVD/module3README.md
@@ -1,14 +1,0 @@
-## API using the VariantValidator LOVD Endpoint to retrieve MANE Select and MANE Select Plus Clinical Data
-
-# INPUT:
-
-Required input to be filled in to fields of the API is selection of genome build and a genomic variant in HGVS format.
-To retrieve the required mane_select transcript data for refseq and ensembl transcript models, additionally configure select_transcripts to mane_select.
-
-# Definition of mane_select in the VariantValidator LOVD Endpoint
-
-mane_select transcript data includes MANE Select and MANE Plus Clinical Data in the VariantValidator LOVD endpoint.
-
-# OUTPUT:
-
-The current output can be downloaded in json format from the reponse body window.
